Add files via upload
Olympic Games are one of the main international event and also a matter of prestige for countries and therefore each country tries to give their best performance during the event. Despite a lot of hard work, many countries/player are unable to perform well during the events and grab medals whereas there are many countries which performs very well in the event and secures many medals.

An Analysis need to be done by each country to evaluate the previous statistics which will detect the mistakes which they have done previously and will also help them in future development.

An analysis can also be done by the host country to find out the mistakes in the arrangements of the Event which will help them in overcoming these mistakes and host the event accurately.

The primary objective is to analyze the large Olympic dataset using Data Analysis to evaluate the evolution of Olympic Games over the years. This analysis will provide detailed and accurate information regarding various factors which leads to the evolution of Olympic Games and improvement of Countries/Players over the time in visual format.

Visualization of the data over various factors will provide us with the statistical view of the various factors which leads to the evolution of the Olympic Games and Improvement in performance of various Countries/Players over the time.